Update: README.md

+++ Virtual Tourist/Controllers/PhotoAlbumViewController.swift
@@ -18,21 +18,20 @@ class PhotoAlbumViewController:UIViewController, UICollectionViewDelegate, UICol
     @IBOutlet weak var newAlbumButton: UIBarButtonItem!
     
     var pin: Pin!
-    
     var dataController: DataController!
     var fetchedResultsController: NSFetchedResultsController<Photo>!
-    var images: [Data]!
     var screenSize: CGRect!
     var navBar: UINavigationBar!
     var navItem: UINavigationItem!
     var checkValid: Bool!
     
+    //MARK: View Override Functions
     override func viewDidLoad() {
         super.viewDidLoad()
         addNavBar()
         setMapView()
         setUpFetchedResultsController()
-        if pin.photos?.count == 0 {
+        if pin.photos?.count == 0 { //If photos do not exist then a new collection will download
             loadNewPictures()
         }
     }
@@ -40,8 +39,6 @@ class PhotoAlbumViewController:UIViewController, UICollectionViewDelegate, UICol
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
         setUpFetchedResultsController()
-        
-        
     }
     
     override func viewDidLayoutSubviews() {
@@ -54,6 +51,7 @@ class PhotoAlbumViewController:UIViewController, UICollectionViewDelegate, UICol
         fetchedResultsController = nil
     }
     
+    //MARK: Adds Navigation bar and back button programatically
     func addNavBar() {
         screenSize = UIScreen.main.bounds
         navBar = UINavigationBar(frame: CGRect(x: 0, y: 0, width: screenSize.width, height: 64))
@@ -69,22 +67,25 @@ class PhotoAlbumViewController:UIViewController, UICollectionViewDelegate, UICol
         self.view.addSubview(navBar)
     }
     
+    //MARK: Dismiss PhotoAlbumViewController
     @objc func backButton() {
-        
         dismiss(animated: true, completion: nil)
     }
     
+    //MARK: Button for new images
     @IBAction func newCollectionButton(_ sender: Any) {
         loadNewPictures()
     }
+    
+    //UI disabling for collection view while images are loading
     func waitForLoad() {
-        
         performUIUpdatesOnMain {
             self.newAlbumButton.isEnabled = !(self.newAlbumButton.isEnabled)
             self.photoCollection.allowsSelection = !self.photoCollection.allowsSelection
         }
     }
     
+    //MARK: Fetch Controller
     func setUpFetchedResultsController() {
         let fetchRequest:NSFetchRequest<Photo> = Photo.fetchRequest()
         let predicate = NSPredicate(format: "pin == %@", pin)
@@ -92,7 +93,7 @@ class PhotoAlbumViewController:UIViewController, UICollectionViewDelegate, UICol
         let sortDescriptor = NSSortDescriptor(key: "creationDate", ascending: false)
         fetchRequest.sortDescriptors = [sortDescriptor]
         
-        fetchedResultsController = NSFetchedResultsController(fetchRequest: fetchRequest, managedObjectContext: dataController.viewContext, sectionNameKeyPath: nil, cacheName: "photo \(pin.creationDate)")
+        fetchedResultsController = NSFetchedResultsController(fetchRequest: fetchRequest, managedObjectContext: dataController.viewContext, sectionNameKeyPath: nil, cacheName: "photo \(pin.creationDate)")//creation date is unique id linking pins and photos
         fetchedResultsController.delegate = self
         do {
             try fetchedResultsController.performFetch()
@@ -100,6 +101,7 @@ class PhotoAlbumViewController:UIViewController, UICollectionViewDelegate, UICol
             fatalError("The Fetch Could Not Be Performed: \(error.localizedDescription)")
         }
     }
+    //MARK: Sets mapview region and disables interaction capabilities
     func setMapView() {
         
         mapView.delegate = self
@@ -121,6 +123,7 @@ class PhotoAlbumViewController:UIViewController, UICollectionViewDelegate, UICol
             self.mapView.setRegion(region, animated: true)
         }
     }
+    //MARK: Creates and adds images to photo object context
     func addPhotoToDataController(image: Data) {
         let photo = Photo(context: dataController.viewContext)
         photo.creationDate = pin.creationDate!
@@ -129,24 +132,29 @@ class PhotoAlbumViewController:UIViewController, UICollectionViewDelegate, UICol
         try? dataController.viewContext.save()
     }
     
+    //MARK: Delete Photo from collection view and viewContext
     func deletePhoto(at indexPath: IndexPath) {
-        
-        let photoToDelete = fetchedResultsController.object(at: indexPath)
-        dataController.viewContext.delete(photoToDelete)
-        try? dataController.viewContext.save()
-        performUIUpdatesOnMain {
-            self.photoCollection.reloadData()
+        if fetchedResultsController.fetchedObjects?.count != 0 {
+            
+            let photoToDelete = fetchedResultsController.object(at: indexPath)
+            dataController.viewContext.delete(photoToDelete)
+            try? dataController.viewContext.save()
+            performUIUpdatesOnMain {
+                self.photoCollection.reloadData()
+            }
         }
     }
     
+    //MARK: Gets new images
     @objc func loadNewPictures() {
         pin.photos = nil
         checkValid = true
         waitForLoad()
         FlickrPhotoDownloader().imagesFromURL(latitude: pin.latitude, longitude: pin.longitude) { (result, error, finished) in
             if error != nil {
-                
+                self.displayError(error?.localizedDescription)
                 self.waitForLoad()
+                self.checkValid = false
             }
             if let photo = result {
                 self.addPhotoToDataController(image: photo)
@@ -154,47 +162,47 @@ class PhotoAlbumViewController:UIViewController, UICollectionViewDelegate, UICol
                 if finished == true {
                     self.checkValid = false
                     self.waitForLoad()
-                    
                 }
             }
             performUIUpdatesOnMain {
                 self.photoCollection.reloadData()
-                
             }
         }
     }
     
+    //MARK: Collection View delegate override
      func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
         if section == 0 {
             photoLabel.text = "No Images"
         }
         if checkValid == true {
             return 15
-        }else{
-            
+        } else {
             return fetchedResultsController.sections?[section].numberOfObjects ?? 0
         }
     }
-
+    //add images to cell
     func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
         
         let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "photoCell", for: indexPath) as! PhotoCell
         
         if (fetchedResultsController.fetchedObjects?.count)! > (indexPath as NSIndexPath).row {
-            
+    
             let aPhoto = fetchedResultsController.object(at: indexPath)
             photoLabel.text = ""
-            cell.photoImage.image = UIImage(data: aPhoto.image!)
+            if let image = aPhoto.image {
+            cell.photoImage.image = UIImage(data: image)
+            }
         }
-
-            return cell
-        }
+        return cell
+    }
     
     func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
             checkValid = false
             deletePhoto(at: indexPath)
         }
 }
+//MARK: MapView Delegate
 extension PhotoAlbumViewController:MKMapViewDelegate {
     
     func mapView(_ mapView: MKMapView, viewFor annotation: MKAnnotation) -> MKAnnotationView? {
@@ -207,8 +215,6 @@ extension PhotoAlbumViewController:MKMapViewDelegate {
             pinView = MKPinAnnotationView(annotation: annotation, reuseIdentifier: reuseId)
             pinView!.canShowCallout = false
             pinView!.pinTintColor = .red
-            
-        
         }
         else {
             pinView!.annotation = annotation
@@ -216,18 +222,12 @@ extension PhotoAlbumViewController:MKMapViewDelegate {
         return pinView
     }
 }
+//MARK: FetchController Delegate
 extension PhotoAlbumViewController:NSFetchedResultsControllerDelegate {
-    func controllerWillChangeContent(_ controller: NSFetchedResultsController<NSFetchRequestResult>) {
-        
-    }
     
-    func controllerDidChangeContent(_ controller: NSFetchedResultsController<NSFetchRequestResult>) {
-        
-    }
     func controller(_ controller: NSFetchedResultsController<NSFetchRequestResult>, didChange anObject: Any, at indexPath: IndexPath?, for type: NSFetchedResultsChangeType, newIndexPath: IndexPath?) {
         switch type {
         case .insert:
-            
            return
         case .delete:
             return
